refactor: move open_image_file to commands module; type the invoke call
