+++ frontend/src/components/Media/MediaInfoPanel.tsx
@@ -218,7 +218,7 @@ export const MediaInfoPanel: React.FC<MediaInfoPanelProps> = ({
                   e.preventDefault();
                   if (currentImage?.path) {
                     try {
-                      await invoke('open_image_file', { path: currentImage.path });
+                      await invoke<void>('open_image_file', { path: currentImage.path });
                     } catch (err) {
                       console.error('Failed to open file:', err);
                     }
